requirements.txt
streamlit==1.41.1
scikit-learn==1.3.2
numpy==1.26.1
openai==2.11.0
python-dotenv==1.0.1